# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dental CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinic/condition-prices:
    get:
      operationId: getConditionPrices
      tags: [dental]
      summary: Get clinic condition prices and ICD-10 codes
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Condition prices map
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConditionPricesResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateConditionPrices
      tags: [dental]
      summary: Update clinic condition prices (owner/admin only)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConditionPricesRequest"
      responses:
        "200":
          description: Condition prices updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConditionPricesResponse"
        "403":
          description: Forbidden - requires owner or admin role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /patients/{id}/teeth:
    get:
      operationId: listTeeth
      tags: [dental]
      summary: Get all tooth records for a patient
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tooth records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeethResponse"
        "404":
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
        code:
          type: string
      required:
        - success
        - error
...
    ToothCondition:
      type: string
      enum:
        - healthy
        - cavity
        - treated
        - crown
        - root_canal
        - implant
        - missing
        - extraction_needed
    ToothRecord:
      type: object
      properties:
        id:
          type: string
        clinicId:
          type: string
        patientId:
          type: string
        toothFdi:
          type: integer
        condition:
          $ref: "#/components/schemas/ToothCondition"
        notes:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          nullable: true
      required:
        - id
        - clinicId
        - patientId
        - toothFdi
        - condition
        - updatedAt
...
    TeethResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          properties:
            teeth:
              type: array
              items:
                $ref: "#/components/schemas/ToothRecord"
          required:
            - teeth
      required:
        - success
        - data
    ConditionPriceEntry:
      type: object
      properties:
        price:
          type: number
        mkb10:
          type: string
      required:
        - price
        - mkb10
    ConditionPricesMap:
      type: object
      additionalProperties:
        $ref: "#/components/schemas/ConditionPriceEntry"
      description: Map of condition name to price and ICD-10 code
    ConditionPricesResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          properties:
            prices:
              $ref: "#/components/schemas/ConditionPricesMap"
          required:
            - prices
      required:
        - success
        - data
    UpdateConditionPriceItem:
      type: object
      properties:
        price:
          type: number
          minimum: 0
        mkb10Code:
          type: string
      required:
        - price
    UpdateConditionPricesRequest:
      type: object
      properties:
        prices:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/UpdateConditionPriceItem"
          description: Map of condition name to price and optional МКБ-10 code
      required:
        - prices
...